+++ src/plugins/vis_builder/public/application/utils/breadcrumbs.ts
@@ -6,14 +6,14 @@
 import { i18n } from '@osd/i18n';
 import { VISUALIZE_ID } from '../../../common';
 
-const defaultEditText = i18n.translate('wizard.editor.defaultEditBreadcrumbText', {
+const defaultEditText = i18n.translate('visBuilder.editor.defaultEditBreadcrumbText', {
   defaultMessage: 'Edit',
 });
 
 export function getVisualizeLandingBreadcrumbs(navigateToApp) {
   return [
     {
-      text: i18n.translate('wizard.listing.breadcrumb', {
+      text: i18n.translate('visBuilder.listing.breadcrumb', {
         defaultMessage: 'Visualize',
       }),
       onClick: () => navigateToApp(VISUALIZE_ID),
@@ -25,7 +25,7 @@ export function getCreateBreadcrumbs(navigateToApp) {
   return [
     ...getVisualizeLandingBreadcrumbs(navigateToApp),
     {
-      text: i18n.translate('wizard.editor.createBreadcrumb', {
+      text: i18n.translate('visBuilder.editor.createBreadcrumb', {
         defaultMessage: 'Create',
       }),
     },